Crate nvtt_rs

nvtt

A rust wrapper around the Nvidia Texture Tools 3 library.

NVTT 3 is a library that can be used to compress image data and files into compressed texture formats, and to handle compressed and uncompressed images.

In NVTT 3, most compression algorithms and image processing algorithms can be accelerated by the GPU. These have CPU fallbacks for GPUs without support for CUDA. CUDA operations can be enabled through the cuda feature.

Dependencies

The NVTT 3 SDK must be installed on the system. A non-standard path to the binaries can be specified via the NVTT_PATH environment variable. A compiler supporting at least C99 and dynamic linking is also required.

Windows

Windows 10 or 11 (64-bit) are required. The Path environment variable must also contain the path to the directory containing nvtt.dll. Note that this must be done manually, it is not done in a standard Nvidia Texture Tools install.

Linux

64-bit only; Ubuntu 16.04+ or a similarly compatible distro is required. libc.so version 6 or higher is required as well.

Limitations

Currently there is no file I/O support, no low-level (nvtt_lowlevel.h) wrapper, and no batch compression.

Using nvtt

// Create a surface
let input = InputFormat::Bgra8Ub {
    data: &[0u8; 16 * 16 * 4],
    unsigned_to_signed:  false,
};
let image = Surface::image(input, 16, 16, 1).unwrap();

// Create the compression context; enable CUDA if possible
let mut context = Context::new();
#[cfg(feature = "cuda")]
if *CUDA_SUPPORTED {
    context.set_cuda_acceleration(true);
}

// Specify compression settings to use; compress to Bc7
let mut compression_options = CompressionOptions::new();
compression_options.set_format(Format::Bc7);

// Specify how to write the compressed data; indicate as sRGB
let mut output_options = OutputOptions::new();
output_options.set_srgb_flag(true);

// Write the DDS header.
let header = context.output_header(
    &image,
    1, // number of mipmaps
    &compression_options,
    &output_options,
).unwrap();

// Compress and write the compressed data.
let bytes = context.compress(
    &image,
    &compression_options,
    &output_options,
).unwrap();

// Bc7 is 1 byte per pixel.
assert_eq!(16 * 16, bytes.len());

Structs

BoundingBox

Bounding box of a Surface. For use with Surface::add_channel() and Surface::create_sub_image().

CUDA_SUPPORTED

Whether CUDA acceleration is supported by this device.

CompressionOptions

Describes the desired compression format and other compression settings.

Context

Compression context.

CubeSurface

One level of a cube map texture.

Filter

A generic filter. Can be either a mipmap filter or a resize filter.

OutputOptions

Holds container type and options specific to the container.

Surface

One level of a 2D or 3D texture with red, green, blue, and alpha channels.

Enums

AlphaMode

Alpha mode. For use with Surface::set_alpha_mode()

Channel

Represents an RGBA channel. For various operations with Surface and CubeSurface.

Container

Container type for encoded data.

CubeFace

Face of a cubemap. For use with CubeSurface::face()

CubeLayout

Specifies how to fold or unfold a cube map from or to a 2D texture. For use with CubeSurface::fold() and CubeSurface::unfold()

EdgeFixup

Affects how certain cube surface processing algorithms work. For use with CubeSurface::cosine_power_filter() and CubeSurface::fast_resample()

Format

Supported block-compression formats, including compressor variants.

InputFormat

Input formats. For use with Surface::image().

Mipmap

A mipmap filter. For use with Surface::build_next_mipmap().

NormalTransform

Specifies a normal transformation. For use with Surface::transform_normals() and Surface::reconstruct_normals().

PixelType

Pixel value types.

Quality

Quality modes.

Resize

A resize filter. For use with Surface::resize_filtered() and Surface::resize_make_square().

RoundMode

Extents rounding mode. For use with Surface::resize_rounded() and Surface::resize_make_square().

SplitInputFormat

Split input formats. For use with Surface::image_split().

SurfaceError

The error type for various Surface operations.

Swizzle

Swizzle order. For use with Surface::swizzle().

TextureType

Texture types. Specifies the dimensionality of a texture.

ToneMapper

Tone mapping functions. For use with Surface::tonemap()

WrapMode

Specifies how to handle coordinates outside the typical image range. For use with Surface::set_wrap_mode()

Functions

version

Return NVTT version.