Struct nvtt_rs::CompressionOptions

pub struct CompressionOptions(_);

Describes the desired compression format and other compression settings.

Implementations

impl CompressionOptions

pub fn new() -> Self

Constructs a new CompressionOptions struct. Sets compression options to the default values.

pub fn reset(&mut self)

Sets compression options to the default values.

pub fn set_format(&mut self, format: Format)

Set desired compression format.

pub fn set_quality(&mut self, quality: Quality)

Set compression quality settings.

pub fn set_color_weights(
&mut self,
red: f32,
green: f32,
blue: f32,
alpha: Option<f32>
)

Set the weights of each color channel used to measure compression error.

The choice for these values is subjective. In most cases uniform color weights (1.0, 1.0, 1.0) work very well. A popular choice is to use the NTSC luma encoding weights (0.2126, 0.7152, 0.0722), but I think that blue contributes to our perception more than 7%. A better choice in my opinion is (3, 4, 2).

Optional Parameters

alpha: Defaults to 1.0

pub unsafe fn set_pixel_format(
&mut self,
bitcount: u32,
rmask: u32,
gmask: u32,
bmask: u32,
amask: u32
)

Describes an RGB/RGBA format using 32-bit masks per channel.

Note that this sets the number of bits per channel to 0.

Safety

See Nvidia SDK documentation for more information.

pub fn set_pixel_type(&mut self, pixel_type: PixelType)

Set pixel type.

These are used for Format::Rgb: they indicate how the output should be interpreted, but do not have any influence over the input. They are ignored for other compression modes.

pub fn set_pitch_alignment(&mut self, pitch_alignment: i32)

Set pitch alignment in bytes.

pub fn set_quantization(
&mut self,
color_dithering: bool,
alpha_dithering: bool,
binary_alpha: bool,
alpha_threshold: Option<i32>
)

Set quantization options.

Optional Parameters

alpha_threshold: Defaults to 127

Warning

Do not enable dithering unless you know what you are doing. Quantization introduces errors. It’s better to let the compressor quantize the result to minimize the error, instead of quantizing the data before handling it to the compressor.

pub fn d3d9_format(&self) -> u32

Translate to and from D3D formats.

Returns 0 if no corresponding format could be found.

For Format::Rgb, this looks at the pixel type and pixel format to determine the corresponding D3D format. For BC6, BC7, and ASTC, this returns a FourCC code: ‘BC6H’ for both unsigned and signed BC6, ‘BC7L’ for BC7, and ‘ASTC’ for all ASTC formats.