Struct nvtt_rs::Context

pub struct Context(_);

Compression context.

Implementations

impl Context

pub fn new() -> Self

Constructs a new compression context.

pub fn set_cuda_acceleration(&mut self, enable: bool)

Available on crate feature cuda only.

Enable/Disable CUDA acceleration; initializes CUDA if not already initialized.

Panics

Panics if CUDA_SUPPORTED is false.

pub fn is_cuda_acceleration_enabled(&self) -> bool

Available on crate feature cuda only.

Check if CUDA acceleration is enabled.

pub fn output_header(
&self,
img: &Surface,
mipmap_count: u32,
compression_options: &CompressionOptions,
output_options: &OutputOptions
) -> Option<Vec<u8>>

Write the Container’s header to a Vec<u8>. Returns Some(Vec<u8>) on success.

pub fn output_header_cube(
&self,
cube: &CubeSurface,
mipmap_count: u32,
compression_options: &CompressionOptions,
output_options: &OutputOptions
) -> Option<Vec<u8>>

Write the Container’s header to a Vec<u8>. Returns Some(Vec<u8>) on success.

pub fn output_header_data(
&self,
tex_type: TextureType,
w: u32,
h: u32,
d: u32,
mipmap_count: u32,
is_normal_map: bool,
compression_options: &CompressionOptions,
output_options: &OutputOptions
) -> Option<Vec<u8>>

Write the Container’s header to a Vec<u8>. Returns Some(Vec<u8>) on success.

pub fn compress(
&self,
img: &Surface,
compression_options: &CompressionOptions,
output_options: &OutputOptions
) -> Option<Vec<u8>>

Compress the Surface and write the compressed data to a Vec<u8>. Returns Some(Vec<u8>) on success.

Examples

let input = InputFormat::Bgra8Ub {
    data: &[0u8; 100 * 100 * 4],
    unsigned_to_signed:  false,
};
let image = Surface::image(input, 100, 100, 1).unwrap();
let context = Context::new();

let mut compression_options = CompressionOptions::new();
// Just leave it as is
compression_options.set_format(Format::Rgba);

let output_options = OutputOptions::new();

// Compress and write the compressed data.
let bytes = context.compress(
    &image,
    &compression_options,
    &output_options,
).unwrap();

assert_eq!(100 * 100 * 4, bytes.len());
for byte in bytes {
    assert_eq!(0, byte);
}

pub fn compress_cube(
&self,
cube: &CubeSurface,
compression_options: &CompressionOptions,
output_options: &OutputOptions
) -> Option<Vec<u8>>

Compress the CubeSurface and write the compressed data to a Vec<u8>. Returns Some(Vec<u8>) on success.

pub fn compress_data(
&self,
w: u32,
h: u32,
d: u32,
rgba: &[f32],
compression_options: &CompressionOptions,
output_options: &OutputOptions
) -> Option<Vec<u8>>

Compress and write data to a Vec<u8>. Returns Some(Vec<u8>) on success.

Panics

Panics if w * h * d < rgba.len().

pub fn estimate_size(
&self,
img: &Surface,
mipmap_count: u32,
compression_options: &CompressionOptions
) -> u32

Returns the total compressed size of mips 0...mipmap_count - 1, without compressing the image.

Note that this does not include the container header, and mips are assumed to be tightly packed.

For instance, call this with mipmap_count = Surface::count_mipmaps() and add the size of the DDS header to get the size of a DDS file with a surface and a full mip chain.

pub fn estimate_size_cube(
&self,
cube: &CubeSurface,
mipmap_count: u32,
compression_options: &CompressionOptions
) -> u32

Returns the total compressed size of mips 0...mipmap_count - 1, without compressing the image.

Note that this does not include the container header, and mips are assumed to be tightly packed.

For instance, call this with mipmap_count = CubeSurface::count_mipmaps() and add the size of the DDS header to get the size of a DDS file with a surface and a full mip chain.

pub fn estimate_size_data(
&self,
w: u32,
h: u32,
d: u32,
mipmap_count: u32,
compression_options: &CompressionOptions
) -> u32

Returns the total compressed size of mips 0...mipmap_count - 1, without compressing the image.

Note that this does not include the container header, and mips are assumed to be tightly packed.