Struct nvtt_rs::CubeSurface

pub struct CubeSurface(_);

One level of a cube map texture.

Contained are six square Surfaces numbered 0 through 5, all with the same size (referred to as the edge length). By convention, these are the +x, -x, +y, -y, +z, and -z faces, in that order, of a cube map in a right-handed coordinate system.

Implementations

impl CubeSurface

pub fn fold(img: &Surface, layout: CubeLayout) -> Self

Creates a CubeSurface from a 2D unfolded Surface in img.

Panics

Panics if img is 3D, or if it does not have a shape which can fold from the specified layout. See also CubeLayout::dim_supported().

pub fn unfold(&self, layout: CubeLayout) -> Surface

Creates a Surface containing an unfolded/flattened representation of the cube surface.

pub fn face(&self, face: CubeFace) -> &Surface

Returns a reference for the given face.

pub fn edge_length(&self) -> u32

Returns the edge length of any of the faces.

Examples

let bytes = [0u8; (10 * 60 * std::mem::size_of::<f32>())];
let input = InputFormat::R32f(&bytes);

let surface = Surface::image(input, 10, 60, 1).unwrap();
let cube_surface = CubeSurface::fold(&surface, CubeLayout::Column);

assert_eq!(10, cube_surface.edge_length());

pub fn count_mipmaps(&self) -> u32

Returns the number of mips that would be in a full mipmap chain starting with this CubeSurface

For instance, a full mip chain for a cube map with 10x10 faces would consist of cube maps with sizes 10x10, 5x5, 2x2, and 1x1, and this function would return 4.

Examples

let bytes = [0u8; (10 * 60 * std::mem::size_of::<f32>())];
let input = InputFormat::R32f(&bytes);

let surface = Surface::image(input, 10, 60, 1).unwrap();
let cube_surface = CubeSurface::fold(&surface, CubeLayout::Column);

assert_eq!(4, cube_surface.count_mipmaps());

pub fn average(&self, channel: Channel) -> f32

Computes an average value for the given channel over the entire sphere.

This takes solid angles into account when producing an average per steradian, so texels near face edges are weighted less than texels near face edges.

No gamma correction is performed, unlike Surface::average().

pub fn range(&self, channel: Channel) -> (f32, f32)

Returns the minimum and maximum values respectively in the given channel.

pub fn clamp(&mut self, channel: Channel, low: f32, high: f32)

Clamps values in the given channel to the range [low, high].

pub fn to_gamma(&mut self, gamma: f32)

Raises channels to the power 1/gamma. gamma=2.2 approximates sRGB-to-linear conversion.

pub fn from_gamma(&mut self, gamma: f32)

Raises channels to the power gamma. gamma=2.2 approximates sRGB-to-linear conversion.

pub fn cosine_power_filter(
&self,
size: u32,
cosine_power: f32,
fixup_method: EdgeFixup
) -> Self

Spherically convolves this CubeSurface with a max(0.0f, cos(theta))^cosinePower kernel, returning a CubeSurface with faces with dimension size x size.

This is useful for generating prefiltered cube maps, as this corresponds to the cosine power used in the Phong reflection model (with energy conservation).

This handles how each cube map texel can have a different solid angle. It also only considers texels for which the value of the kernel (without normalization) is at least 0.001.

Panics

Panics is size is 0.

pub fn fast_resample(&self, size: u32, fixup_method: EdgeFixup) -> Self

Produces a resized version of this CubeSurface using nearest-neighbor sampling.

Panics

Panics is size is 0.