Struct nvtt_rs::Surface

#[repr(transparent)]
pub struct Surface(_);

One level of a 2D or 3D texture with red, green, blue, and alpha channels.

Surfaces store some additional properties, such as their width, height, depth, wrap mode, alpha mode, and whether they represent a normal map.

Texture data is stored non-interleaved; that is, all the red channel data is stored first, followed by the green channel data, and so on.

Performance Notes

Surfaces use reference-counted pointers to image data. This means that multiple Surfaces can reference the same data. This is handled automatically by NVTT’s image processing routines. For instance, after the following piece of code,

let s1 = Surface::image(...);
let mut s2 = s1.clone();

surfaces s1 and s2 will have the same data() pointer. Cloning the underlying data is handled automatically: for instance, after

s2.to_srgb();

s2 will have a new data() pointer, and s1 will be unchanged.

Implementations

impl Surface

pub fn image(
input: InputFormat<'_>,
w: u32,
h: u32,
d: u32
) -> Result<Self, SurfaceError>

Constructs a Surface from an uncompressed byte slice of all channels. Bytes should be in native endianness.

Data must be stored in [c, z, y, x] order. That is, channel 0’s data is first, then channel 1’s data, and so on. More specifically, the value of a channel c at texel (x, y, z) is at index ((c * d + z) * h + y) * w + x. Where w, h, and d are input width, height, and depth of the image respectively.

If the input slice is not capable of holding w * h * d texels in the specified format, this will return SurfaceError::InvalidDimensions. If the slice is longer than necessary, extra bytes remain unread.

Formats without four R, G, B, and A channels (currently only InputFormat::R32f) will have non-existent channels filled with the color [0, 0, 0, 0].

Panics

Panics if any of w, h, or d are 0.

Examples

let input = InputFormat::Bgra8Ub {
    data: &[0, 255, 255, 0],
    unsigned_to_signed:  false,
};
let surface = Surface::image(input, 1, 1, 1).unwrap();
assert_relative_eq!(1.0, surface.channel(Channel::R)[0]);
assert_relative_eq!(1.0, surface.channel(Channel::G)[0]);
assert_relative_eq!(0.0, surface.channel(Channel::B)[0]);
assert_relative_eq!(0.0, surface.channel(Channel::A)[0]);

let input = InputFormat::Bgra8Ub {
    data: &[0, 255, 255, 0],
    unsigned_to_signed:  true,
};
let surface = Surface::image(input, 1, 1, 1).unwrap();
assert_relative_eq!(1.0, surface.channel(Channel::R)[0]);
assert_relative_eq!(1.0, surface.channel(Channel::G)[0]);
assert_relative_eq!(-1.0, surface.channel(Channel::B)[0]);
assert_relative_eq!(-1.0, surface.channel(Channel::A)[0]);

let r_bytes = 3.0_f32.to_ne_bytes();
let input = InputFormat::R32f(&r_bytes);
let surface = Surface::image(input, 1, 1, 1).unwrap();
assert_relative_eq!(3.0, surface.channel(Channel::R)[0]);
assert_relative_eq!(0.0, surface.channel(Channel::G)[0]);
assert_relative_eq!(0.0, surface.channel(Channel::B)[0]);
assert_relative_eq!(0.0, surface.channel(Channel::A)[0]);

let input = InputFormat::Bgra8Ub {
    data: &[0, 0, 0, 0],
    unsigned_to_signed:  false,
};
let surface = Surface::image(input, 100, 100, 100);
assert!(surface.is_err());

pub fn image_split(
input: SplitInputFormat<'_>,
w: u32,
h: u32,
d: u32
) -> Result<Self, SurfaceError>

Constructs a Surface from uncompressed byte slices of all channels. Bytes should be in native endianness.

Data must be stored in [z, y, x] order. More specifically, the value of any channel at texel (x, y, z) is at index ((d + z) * h + y) * w + x. Where w, h, and d are input width, height, and depth of the image respectively.

If any input slice is not capable of holding w * h * d texels in the specified format, this will return SurfaceError::InvalidDimensions. If the slice is longer than necessary, extra bytes remain unread.

Formats without four R, G, B, and A channels (currently only SplitInputFormat::R32f) will have non-existent channels filled with the color [0, 0, 0, 0].

Panics

Panics if any of w, h, or d are 0.

Examples

let input = SplitInputFormat::Bgra8Ub {
    b: &[0],
    g: &[255],
    r: &[255],
    a: &[0],
};
let surface = Surface::image_split(input, 1, 1, 1).unwrap();
assert_relative_eq!(1.0, surface.channel(Channel::R)[0]);
assert_relative_eq!(1.0, surface.channel(Channel::G)[0]);
assert_relative_eq!(0.0, surface.channel(Channel::B)[0]);
assert_relative_eq!(0.0, surface.channel(Channel::A)[0]);

let r_bytes = 3.0_f32.to_ne_bytes();
let input = SplitInputFormat::R32f(&r_bytes);
let surface = Surface::image_split(input, 1, 1, 1).unwrap();
assert_relative_eq!(3.0, surface.channel(Channel::R)[0]);
assert_relative_eq!(0.0, surface.channel(Channel::G)[0]);
assert_relative_eq!(0.0, surface.channel(Channel::B)[0]);
assert_relative_eq!(0.0, surface.channel(Channel::A)[0]);

let input = SplitInputFormat::Bgra8Ub {
    b: &[0],
    g: &[0],
    r: &[0],
    a: &[0],
};
let surface = Surface::image_split(input, 100, 100, 100);
assert!(surface.is_err());

pub fn width(&self) -> u32

Returns the width (X size) of the surface in pixels.

pub fn height(&self) -> u32

Returns the height (Y size) of the surface in pixels.

pub fn depth(&self) -> u32

Returns the depth (Z size) of the surface in pixels.

pub fn on_cpu(&self) -> bool

Available on crate feature cuda only.

Returns true if data is currently held on the CPU. This is the same as checking if Surface::gpu_data_ptr() is null.

#[cfg(feature = "cuda")]
if *CUDA_SUPPORTED {
    let input_format = InputFormat::Bgra8Ub {
        data: &[255, 255, 255, 255],
        unsigned_to_signed: false,
    };
    let mut surface = Surface::image(input_format, 1, 1, 1).unwrap();
    assert!(surface.on_cpu());
    assert!(surface.gpu_data_ptr().is_null());

    surface.to_gpu();
    assert!(!surface.on_cpu());
    assert!(!surface.gpu_data_ptr().is_null());

    surface.to_cpu();
    assert!(surface.on_cpu());
    assert!(surface.gpu_data_ptr().is_null());
}

pub fn on_gpu(&self) -> bool

Available on crate feature cuda only.

Returns true if data is currently held to the GPU. This is the same as checking if Surface::gpu_data_ptr() is not null.

#[cfg(feature = "cuda")]
if *CUDA_SUPPORTED {
    let input_format = InputFormat::Bgra8Ub {
        data: &[255, 255, 255, 255],
        unsigned_to_signed: false,
    };
    let mut surface = Surface::image(input_format, 1, 1, 1).unwrap();
    assert!(!surface.on_gpu());
    assert!(surface.gpu_data_ptr().is_null());

    surface.to_gpu();
    assert!(surface.on_gpu());
    assert!(!surface.gpu_data_ptr().is_null());

    surface.to_cpu();
    assert!(!surface.on_gpu());
    assert!(surface.gpu_data_ptr().is_null());
}

pub fn to_gpu(&mut self)

Available on crate feature cuda only.

Copies data from CPU to GPU, enabling CUDA for all subsequent operations. Does nothing if data is already on the GPU.

Panics

Panics if CUDA is not supported by this device.

Examples

#[cfg(feature = "cuda")]
if *CUDA_SUPPORTED {
    let input_format = InputFormat::Bgra8Ub {
        data: &[255, 255, 255, 255],
        unsigned_to_signed: false,
    };
    let mut surface = Surface::image(input_format, 1, 1, 1).unwrap();
    surface.to_gpu();
    assert!(surface.on_gpu());
}

if !*CUDA_SUPPORTED {
    let input_format = InputFormat::Bgra8Ub {
        data: &[255, 255, 255, 255],
        unsigned_to_signed: false,
    };
    let mut surface = Surface::image(input_format, 1, 1, 1).unwrap();
    // Panics
    surface.to_gpu();
}

pub fn to_cpu(&mut self)

Available on crate feature cuda only.

Copies data from GPU to CPU, disabling CUDA operations. Does nothing if data is already on the CPU.

Examples

#[cfg(feature = "cuda")]
if *CUDA_SUPPORTED {
    let input_format = InputFormat::Bgra8Ub {
        data: &[255, 255, 255, 255],
        unsigned_to_signed: false,
    };
    let mut surface = Surface::image(input_format, 1, 1, 1).unwrap();
    surface.to_gpu();
    assert!(surface.on_gpu());

    surface.to_cpu();
    assert!(surface.on_cpu());
}

if !*CUDA_SUPPORTED {
    let input_format = InputFormat::Bgra8Ub {
        data: &[255, 255, 255, 255],
        unsigned_to_signed: false,
    };
    let mut surface = Surface::image(input_format, 1, 1, 1).unwrap();
    assert!(surface.on_cpu());

    surface.to_cpu();
    assert!(surface.on_cpu());
}

pub fn gpu_data_ptr(&self) -> *const f32

Available on crate feature cuda only.

Returns a CUDA pointer to image data on GPU, with the same layout as Surface::data(). If data is not currently stored on the GPU, this is a null pointer.

Safety

It is undefined behaviour to derefence this pointer on the CPU, as it is not a CPU pointer.

#[cfg(feature = "cuda")]
if *CUDA_SUPPORTED {
    let input_format = InputFormat::Bgra8Ub {
        data: &[255, 255, 255, 255],
        unsigned_to_signed: false,
    };
    let mut surface = Surface::image(input_format, 1, 1, 1).unwrap();
    surface.to_gpu();

    assert!(!surface.gpu_data_ptr().is_null());

    // Undefined behaviour
    unsafe {
        let x = *surface.gpu_data_ptr();
    }
}

pub fn texel(&self, channel: Channel, x: u32, y: u32, z: u32) -> f32

Returns the value of the channel for the texel at (x, y, z). This is just a wrapper around Surface::data() for easier indexing. See Surface::data() for more details.

Panics

Panics if (x, y, z) is out of bounds.

pub fn texel_mut(&mut self, channel: Channel, x: u32, y: u32, z: u32) -> &mut f32

Returns a mutable reference to the value of the channel for the texel at (x, y, z). This is just a wrapper around Surface::data_mut() for easier indexing. See Surface::data_mut() for more details.

Panics

Panics if (x, y, z) is out of bounds.

pub fn data(&self) -> &[f32]

Returns image data as a slice in [c, z, y, x] order. That is, the value of a channel c at texel (x, y, z) is at index ((c * d + z) * h + y) * w + x.

Notes

If data is on the GPU, this will perform a GPU-CPU copy. CUDA will remain enabled.

Examples

let r_bytes = 1.0_f32.to_ne_bytes();
let input = InputFormat::R32f(&r_bytes);
let mut surface = Surface::image(input, 1, 1, 1).unwrap();

#[cfg(feature = "cuda")]
{
    if *CUDA_SUPPORTED {
        surface.to_gpu();
        assert!(surface.on_gpu());
    } else {
        assert!(surface.on_cpu());
    }
}

// Incurs a GPU-CPU copy
assert_eq!(4, surface.data().len());

// Incurs another GPU-CPU copy
let x = surface.data()[0];
assert_relative_eq!(1.0, x);
println!("{}", x);

// Remains on GPU
#[cfg(feature = "cuda")]
{
    if *CUDA_SUPPORTED {
        assert!(surface.on_gpu());
    } else {
        assert!(surface.on_cpu());
    }
}

pub fn data_mut(&mut self) -> &mut [f32]

Returns image data as a mutable slice in [c, z, y, x] order. That is, the value of a channel c at texel (x, y, z) is at index ((c * d + z) * h + y) * w + x.

Notes

If data is on the GPU, this will perform a GPU-CPU copy. CUDA will be disabled, however it can be re-enabled with Surface::to_gpu().

Examples

let r_bytes = 1.0_f32.to_ne_bytes();
let input = InputFormat::R32f(&r_bytes);
let mut surface = Surface::image(input, 1, 1, 1).unwrap();

#[cfg(feature = "cuda")]
{
    if *CUDA_SUPPORTED {
        surface.to_gpu();
        assert!(surface.on_gpu());
    } else {
        assert!(surface.on_cpu());
    }
}

// Moves to CPU
surface.data_mut()[0] = 0.0;

#[cfg(feature = "cuda")]
assert!(surface.on_cpu());

pub fn channel(&self, channel: Channel) -> &[f32]

Returns a slice of image data for channel in [z, y, x] order. That is, the value of texel (x, y, z) is at index ((d + z) * h + y) * w + x.

Notes

If data is on the GPU, this will perform a GPU-CPU copy. CUDA will remain enabled.

Examples

let r_bytes = 1.0_f32.to_ne_bytes();
let input = InputFormat::R32f(&r_bytes);
let mut surface = Surface::image(input, 1, 1, 1).unwrap();

#[cfg(feature = "cuda")]
{
    if *CUDA_SUPPORTED {
        surface.to_gpu();
        assert!(surface.on_gpu());
    } else {
        assert!(surface.on_cpu());
    }
}

// Each call incurs a GPU-CPU copy
assert_eq!(1, surface.channel(Channel::R).len());
assert_eq!(1, surface.channel(Channel::G).len());
assert_eq!(1, surface.channel(Channel::B).len());
assert_eq!(1, surface.channel(Channel::A).len());

// Incurs another GPU-CPU copy
let x = surface.channel(Channel::R)[0];
println!("{}", x);
assert_relative_eq!(1.0, x);

// Remains on GPU
#[cfg(feature = "cuda")]
{
    if *CUDA_SUPPORTED {
        surface.to_gpu();
        assert!(surface.on_gpu());
    } else {
        assert!(surface.on_cpu());
    }
}

pub fn channel_mut(&mut self, channel: Channel) -> &mut [f32]

Returns a mutable slice of image data for channel in [z, y, x] order. That is, the value of texel (x, y, z) is at index ((d + z) * h + y) * w + x.

Notes

If data is on the GPU, this will perform a GPU-CPU copy. CUDA will be disabled, however it can be re-enabled with Surface::to_gpu().

Examples

let r_bytes = 1.0_f32.to_ne_bytes();
let input = InputFormat::R32f(&r_bytes);
let mut surface = Surface::image(input, 1, 1, 1).unwrap();

#[cfg(feature = "cuda")]
{
    if *CUDA_SUPPORTED {
        surface.to_gpu();
        assert!(surface.on_gpu());
    } else {
        assert!(surface.on_cpu());
    }
}

// Moves to CPU
surface.channel_mut(Channel::R)[0] = 0.0;

#[cfg(feature = "cuda")]
assert!(surface.on_cpu());

pub fn tex_type(&self) -> TextureType

Returns TextureType::D2 if depth == 1, and TextureType::D3 otherwise.

pub fn wrap_mode(&self) -> WrapMode

Returns the WrapMode of this surface. Note that WrapMode can affect operations such as Surface::build_next_mipmap() or Surface::convolve().

pub fn alpha_mode(&self) -> AlphaMode

Returns the AlphaMode of this surface. This is for output headers, e.g. for usage in Context::output_header. It does not affect calculations.

pub fn is_normal_map(&self) -> bool

Returns the true if this surface represents a normal map. This is for output headers, e.g. for usage in Context::output_header. It does not affect calculations.

pub fn set_wrap_mode(&mut self, wrap_mode: WrapMode)

Set the WrapMode of this surface. Note that WrapMode can affect operations such as Surface::build_next_mipmap() or Surface::convolve().

pub fn set_alpha_mode(&mut self, alpha_mode: AlphaMode)

Set the AlphaMode of this surface. This is for output headers, e.g. for usage in Context::output_header. It does not affect calculations.

pub fn set_normal_map(&mut self, is_normal_map: bool)

Set whether this surface represents a normal map. This is for output headers, e.g. for usage in Context::output_header. It does not affect calculations.

pub fn flip_x(&mut self)

Flip along the X axis.

pub fn flip_y(&mut self)

Flip along the Y axis.

pub fn flip_z(&mut self)

Flip along the Z axis.

pub fn copy_channel(
&mut self,
other: &Self,
src_channel: Channel,
dst_channel: Channel
) -> Result<(), SurfaceError>

Copy src_channel from other to dst_channel of this surface. If channels are not the same length, this returns SurfaceError::InvalidDimensions

pub fn add_channel(
&mut self,
other: &Self,
src_channel: Channel,
dst_channel: Channel,
scale: f32
) -> Result<(), SurfaceError>

Add src_channel from other multiplied by scale to dst_channel of this surface. If channels are not the same length, this returns SurfaceError::InvalidDimensions

pub fn bounds(&self) -> BoundingBox

Returns the BoundingBox of this surface. min_x/min_y/min_z are all 0, while max_x, min_y, and min_z are Surface::width(), Surface::height(), and Surface::depth() respectively.

pub fn copy(&mut self, other: &Self, src: BoundingBox, dst: BoundingBox)

Copies the bounding box src of other to the bounding box dst of this surface.

Panics

Panics if src and dst are not the same shape, or one of the bounding boxes is out of bounds.

pub fn create_sub_image(&self, bounds: BoundingBox) -> Self

Creates a sub image from the subset bounds of this surface.

Panics

Panics if this surface does not contain bounds.

pub fn count_mipmaps(&self) -> u32

Returns the number of mipmaps in a full mipmap chain. Each mip is half the size of the previous, rounding down, until and including a 1x1 mip.

For instance, a 8x5 surface has mipmaps of size 8x5 (mip 0), 4x2 (mip 1), 2x1 (mip 2), and 1x1 (mip 3), so Surface::count_mipmaps() returns 4. A 7x3 surface has mipmaps of size 7x3, 3x1, and 1x1, so Surface::count_mipmaps() returns 3.

Same as Surface::count_mipmaps_until() with min_size == 1

Examples

let bytes = [0u8; (8 * 5 * std::mem::size_of::<f32>())];
let input = InputFormat::R32f(&bytes);

let surface = Surface::image(input, 8, 5, 1).unwrap();
assert_eq!(4, surface.count_mipmaps())

let bytes = [0u8; (7 * 3 * std::mem::size_of::<f32>())];
let input = InputFormat::R32f(&bytes);

let surface = Surface::image(input, 7, 3, 1).unwrap();
assert_eq!(3, surface.count_mipmaps())

pub fn count_mipmaps_until(&self, min_size: u32) -> u32

Returns the number of mipmaps in a mipmap chain, stopping when Surface::can_make_next_mipmap() returns false.

That is, it stops when a 1x1x1 mip is reached if min_size == 1 (in which case it is the same as Surface::count_mipmaps()), or stops when the width and height are less than min_size and the depth is 1.

pub fn can_make_next_mipmap(&self, min_size: u32) -> bool

Returns whether the surface would have a next mip in a mip chain with minimum size min_size.

That is, it returns false if this surface has size 1x1x1, or if the width and height are less than min_size and the depth is 1.

pub fn build_next_mipmap(&mut self, filter: Filter<Mipmap>, min_size: u32) -> bool

Resizes this surface to create the next mip in a mipmap chain.

Returns false iff the next mip would have been smaller than min_size (signaling the end of the mipmap chain).

Examples

let bytes = [0u8; (32 * 32 * std::mem::size_of::<f32>())];
let input = InputFormat::R32f(&bytes);
let mut surface = Surface::image(input, 32, 32, 1).unwrap();

let mipmap_count = surface.count_mipmaps();
let mut mipmaps = vec![surface.clone()];
while surface.build_next_mipmap(Filter::mipmap_box(), 1) {
    mipmaps.push(surface.clone());
}

assert_eq!(mipmap_count, mipmaps.len() as u32);

assert_eq!(mipmaps[0].width(), mipmaps[0].height());
assert_eq!(mipmaps[1].width(), mipmaps[1].height());
assert_eq!(mipmaps[2].width(), mipmaps[2].height());
assert_eq!(mipmaps[3].width(), mipmaps[3].height());
assert_eq!(mipmaps[4].width(), mipmaps[4].height());
assert_eq!(mipmaps[5].width(), mipmaps[5].height());

assert_eq!(32, mipmaps[0].width());
assert_eq!(16, mipmaps[1].width());
assert_eq!(8, mipmaps[2].width());
assert_eq!(4, mipmaps[3].width());
assert_eq!(2, mipmaps[4].width());
assert_eq!(1, mipmaps[5].width());

pub fn build_next_mipmap_color(&mut self, rgba: [f32; 4]) -> bool

Replaces this surface with a surface the size of the next mip in a mip chain (half the width and height), but with each channel cleared to a constant value.

Examples

let bytes = [0u8; (32 * 32 * std::mem::size_of::<f32>())];
let input = InputFormat::R32f(&bytes);
let mut surface = Surface::image(input, 32, 32, 1).unwrap();

let rgba = [0.25, 0.5, 0.75, 1.0];
surface.build_next_mipmap_color(rgba);

assert_eq!(16, surface.width());
assert_eq!(16, surface.height());

for x in 0..16 {
    for y in 0..16 {
        assert_relative_eq!(0.25, surface.texel(Channel::R, x, y, 0));
        assert_relative_eq!(0.5,  surface.texel(Channel::G, x, y, 0));
        assert_relative_eq!(0.75, surface.texel(Channel::B, x, y, 0));
        assert_relative_eq!(1.0,  surface.texel(Channel::A, x, y, 0));
    }
}

pub fn fill(&mut self, rgba: [f32; 4])

Sets all texels in the surface to a solid color.

Examples

let bytes = [0u8; (32 * 32 * std::mem::size_of::<f32>())];
let input = InputFormat::R32f(&bytes);
let mut surface = Surface::image(input, 32, 32, 1).unwrap();

let rgba = [0.0, 0.33, 0.66, 1.0];
surface.fill(rgba);

assert_eq!(32, surface.width());
assert_eq!(32, surface.height());

for x in 0..32 {
    for y in 0..32 {
        assert_relative_eq!(0.0,  surface.texel(Channel::R, x, y, 0));
        assert_relative_eq!(0.33, surface.texel(Channel::G, x, y, 0));
        assert_relative_eq!(0.66, surface.texel(Channel::B, x, y, 0));
        assert_relative_eq!(1.0,  surface.texel(Channel::A, x, y, 0));
    }
}

pub fn set_border(&mut self, rgba: [f32; 4])

Sets all texels on the border of the surface to a solid color.

Examples

let bytes = [0u8; (4 * 4 * std::mem::size_of::<f32>())];
let input = InputFormat::R32f(&bytes);
let mut surface = Surface::image(input, 4, 4, 1).unwrap();

let rgba = [0.25, 0.5, 0.75, 1.0];
surface.set_border(rgba);

for x in 0..4 {
    for y in 0..4 {
        // On the border
        if x == 0 || y == 0 || x == 3 || y == 3 {
            assert_relative_eq!(0.25, surface.texel(Channel::R, x, y, 0));
            assert_relative_eq!(0.5,  surface.texel(Channel::G, x, y, 0));
            assert_relative_eq!(0.75, surface.texel(Channel::B, x, y, 0));
            assert_relative_eq!(1.0,  surface.texel(Channel::A, x, y, 0));
        }

        // Not on the border
        else {
            assert_relative_eq!(f32::from_bits(0), surface.texel(Channel::R, x, y, 0));
            assert_relative_eq!(f32::from_bits(0), surface.texel(Channel::G, x, y, 0));
            assert_relative_eq!(f32::from_bits(0), surface.texel(Channel::B, x, y, 0));
            assert_relative_eq!(f32::from_bits(0), surface.texel(Channel::A, x, y, 0));
        }
    }
}

pub fn set_atlas_border(&mut self, w: u32, h: u32, rgba: [f32; 4])

Draws borders of a given color around each w x h tile contained within the surface, starting from the (0, 0) corner. In case the surface size is not divisible by the tile size, borders are not drawn for tiles crossing the surface boundary

Panics

Panics if w or h are 0.

pub fn resize_filtered(&mut self, w: u32, h: u32, d: u32, filter: Filter<Resize>)

Resizes this surface to have size w x h x d using a given filter.

Panics

Panics if w, h, or d are 0

pub fn resize_rounded(
&mut self,
max_extent: u32,
mode: RoundMode,
filter: Filter<Resize>
)

Resizes this surface so that its largest side has length max_extent, subject to a rounding mode and a filter.

Panics

Panics if max_extent is 0.

pub fn resize_make_square(
&mut self,
max_extent: u32,
mode: RoundMode,
filter: Filter<Resize>
)

Resizes this surface so that its largest side has length max_extent and the result is square or cubical. Uses a rounding mode and a filter.

Panics

Panics if max_extent is 0.

pub fn canvas_resize(&mut self, w: u32, h: u32, d: u32)

Crops or expands this surface from the (0, 0, 0) corner, with any new values cleared to 0.

Panics

Panics if any of w, h, or d are 0.

Examples

let bytes = [0u8; (4 * 4 * std::mem::size_of::<f32>())];
let input = InputFormat::R32f(&bytes);
let mut surface = Surface::image(input, 4, 4, 1).unwrap();

let rgba = [1.0, 1.0, 1.0, 1.0];
surface.fill(rgba);
surface.canvas_resize(8, 8, 1);

for x in 0..8 {
    for y in 0..8 {
        // Old region
        if x < 4 && y < 4 {
            assert_relative_eq!(1.0, surface.texel(Channel::R, x, y, 0));
            assert_relative_eq!(1.0, surface.texel(Channel::G, x, y, 0));
            assert_relative_eq!(1.0, surface.texel(Channel::B, x, y, 0));
            assert_relative_eq!(1.0, surface.texel(Channel::A, x, y, 0));
        }

        // New region
        else {
            assert_relative_eq!(0.0, surface.texel(Channel::R, x, y, 0));
            assert_relative_eq!(0.0, surface.texel(Channel::G, x, y, 0));
            assert_relative_eq!(0.0, surface.texel(Channel::B, x, y, 0));
            assert_relative_eq!(0.0, surface.texel(Channel::A, x, y, 0));
        }
    }
}

pub fn premultiply_alpha(&mut self)

Converts to premultiplied alpha, replacing (r, g, b, a) with (ar, ag, ab, a).

Examples

let input = InputFormat::Bgra8Ub {
    data: &[255, 255, 255, 0],
    unsigned_to_signed: true,
};
let mut surface = Surface::image(input, 1, 1, 1).unwrap();
assert_relative_eq!(1.0,  surface.texel(Channel::R, 0, 0, 0));
assert_relative_eq!(1.0,  surface.texel(Channel::G, 0, 0, 0));
assert_relative_eq!(1.0,  surface.texel(Channel::B, 0, 0, 0));
assert_relative_eq!(-1.0, surface.texel(Channel::A, 0, 0, 0));

surface.premultiply_alpha();
assert_relative_eq!(-1.0, surface.texel(Channel::R, 0, 0, 0));
assert_relative_eq!(-1.0, surface.texel(Channel::G, 0, 0, 0));
assert_relative_eq!(-1.0, surface.texel(Channel::B, 0, 0, 0));
assert_relative_eq!(-1.0, surface.texel(Channel::A, 0, 0, 0));

pub fn demultiply_alpha(&mut self, epsilon: Option<f32>)

Converts from premultiplied to unpremultiplied alpha, with special handling around zero alpha values.

When abs(a) >= epsilon, the result is the same as dividing the RGB channels by the alpha channel. Otherwise, this function divides the RGB channels by epsilon * sign(a), since the result of unpremultiplying a fully transparent color is undefined.

Optional Parameters

epsilon: Defaults to 1e-12

Panics

Panics if epsilon is 0.

Examples

let input = InputFormat::Bgra8Ub {
    data: &[0, 0, 0, 0],
    unsigned_to_signed: true,
};
let mut surface = Surface::image(input, 1, 1, 1).unwrap();
assert_relative_eq!(-1.0, surface.texel(Channel::R, 0, 0, 0));
assert_relative_eq!(-1.0, surface.texel(Channel::G, 0, 0, 0));
assert_relative_eq!(-1.0, surface.texel(Channel::B, 0, 0, 0));
assert_relative_eq!(-1.0, surface.texel(Channel::A, 0, 0, 0));

surface.demultiply_alpha(None);
assert_relative_eq!(1.0,  surface.texel(Channel::R, 0, 0, 0));
assert_relative_eq!(1.0,  surface.texel(Channel::G, 0, 0, 0));
assert_relative_eq!(1.0,  surface.texel(Channel::B, 0, 0, 0));
assert_relative_eq!(-1.0, surface.texel(Channel::A, 0, 0, 0));

pub fn to_grey_scale(&mut self, rgba_scale: [f32; 4])

Sets channels to the result of converting to grayscale, with customizable channel weights.

For instance, this can be used to give green a higher weight than red or blue when computing luminance. This function will normalize the different scales so they sum to 1, so e.g. (2, 4, 1, 0) are valid scales. The greyscale value is then computed using

grey = r * rgba_scale[0] + g * rgba_scale[1] + b * rgba_scale[2] + a * rgba_scale[3]

and then all channels (including alpha) are set to grey.

Examples

let input = InputFormat::Bgra8Ub {
    data: &[53, 234, 26, 158],
    unsigned_to_signed: false,
};
let mut surface = Surface::image(input, 1, 1, 1).unwrap();
surface.to_grey_scale([2.0, 4.0, 1.0, 0.0]);

let r = surface.texel(Channel::R, 0, 0, 0);
let g = surface.texel(Channel::G, 0, 0, 0);
let b = surface.texel(Channel::B, 0, 0, 0);
let a = surface.texel(Channel::A, 0, 0, 0);

assert_relative_eq!(r, g);
assert_relative_eq!(g, b);
assert_relative_eq!(b, a);

pub fn from_gamma(&mut self, gamma: f32)

Raises RGB channels to the power gamma. gamma=2.2 approximates sRGB-to-linear conversion.

pub fn channel_from_gamma(&mut self, channel: Channel, gamma: f32)

Raises the given channel to the power gamma.

pub fn to_gamma(&mut self, gamma: f32)

Raises RGB channels to the power 1/gamma. gamma=2.2 approximates sRGB-to-linear conversion.

pub fn channel_to_gamma(&mut self, channel: Channel, gamma: f32)

Raises the given channel to the power 1/gamma.

pub fn to_srgb(&mut self)

Applies the linear-to-sRGB transfer function to RGB channels.

This transfer function replaces each value x with

if x is NaN or x <= 0.0f, 0.0f
if x <= 0.0031308f, 12.92f * x
if x <= 1.0f, powf(x, 0.41666f) * 1.055f - 0.055f
otherwise, 1.0f

pub fn from_srgb(&mut self)

Applies the sRGB-to-linear transfer function to RGB channels.

This transfer function replaces each value x with

if x < 0.0f, 0.0f
if x < 0.04045f, x / 12.92f
if x < 1.0f, powf((x + 0.055f)/1.055f, 2.4f)
otherwise, 1.0f

pub fn to_xenon_srgb(&mut self)

Converts RGB channels from linear to a piecewise linear sRGB approximation.

This transfer function replaces each value x with

if x < 0, 0.0f
if x < 1/16, 4.0f * x
if x < 1/8, 0.25f + 2.0f * (x - 0.0625f)
if x < 1/2, 0.375f + (x - 0.125f)
if x < 1, 0.75f + 0.5f * (x - 0.5f)
otherwise, 1.0f

pub fn to_rgbm(&mut self, range: Option<f32>, threshold: Option<f32>)

Produces an LDR Red, Green, Blue, Magnitude encoding of the HDR RGB channels. See Surface::from_rgbm() for the storage method. This uses an iterative compression approach to reduce the error with regard to decoding.

Optional Parameters

range: Defaults to 1.0

threshold: Defaults to 0.25

pub fn from_rgbm(&mut self, range: Option<f32>, threshold: Option<f32>)

Produces HDR (r, g, b, 1) values from an LDR (red, green, blue, magnitude) storage method.

HDR values are reconstructed as follows: First, the magnitude M is reconstructed from the alpha channel using M = a * (range - threshold) + threshold. Then the red, green, and blue channels are multiplied by M.

Optional Parameters

range: Defaults to 1.0

threshold: Defaults to 0.25

pub fn tonemap(&mut self, tm: ToneMapper)

Applies an HDR-to-LDR tone mapper.

pub unsafe fn to_rgbe(&mut self, mantissa_bits: u32, exponent_bits: u32)

Produces a shared-exponent Red, Green, Blue, Exponent encoding of the HDR RGB channels, such as R9G9B9E5.

See Surface::from_rgbe() for the storage method. This uses an iterative compression approach to reduce the error with regard to decoding.

Safety

The Nvidia SDK does not declare what a safe value for mantissa_bits/exponent_bits may be. If they are too large or small, it will cause undefined behaviour. Assuming the operation has not changed much since the original open sourced implementation, then mantissa_bits/exponent_bits cannot be larger than 32, as they will both be casted to a c_int and then perform the bitshifts 1 << mantissa_bits and 1 << exponent_bits. No current version lower/upper bounds are formally stated.

Only the values mantissa_bits = 9 and exponent_bits = 5 are mentioned, however a far broader range of values is likely safe.

pub unsafe fn from_rgbe(&mut self, mantissa_bits: u32, exponent_bits: u32)

Produces HDR (r, g, b, 1) values from an LDR (red, green, blue, exponent) storage method.

HDR values are reconstructed as follows: R, G, B, and E are first converted from UNORM floats to integers by multiplying RGB by (1 << mantissaBits) - 1 and E by (1 << exponentBits) - 1. E stores a scaling factor as a power of 2, which is reconstructed using scale = 2^(E - ((1 << (exponentBits - 1)) - 1) - mantissaBits). R, G, and B are then multiplied by scale.

Safety

The Nvidia SDK does not declare what a safe value for mantissa_bits/exponent_bits may be. If they are too large or small, it will cause undefined behaviour. Assuming the operation has not changed much since the original open sourced implementation, then mantissa_bits/exponent_bits cannot be larger than 32, as they will both be casted to a c_int and then perform the bitshifts 1 << mantissa_bits and 1 << exponent_bits. No current version lower/upper bounds are formally stated.

Only the values mantissa_bits = 9 and exponent_bits = 5 are mentioned, however a far broader range of values is likely safe.

pub fn to_ycocg(&mut self)

Converts from (r, g, b, -) colors to (Co, Cg, 1, Y) colors.

This is useful for formats that use chroma subsampling.

Y is in the range [0, 1], while Co and Cg are in the range [-1, 1].

The RGB-to-YCoCg formula used is

Y  = (2g + r + b)/4
Co = r - b
Cg = (2g - r - b)/2

pub unsafe fn block_scale_cocg(&mut self, bits: Option<u32>)

Stores per-block YCoCg scaling information for potentially better 4-channel compression of YCoCg data.

For each 4x4 block, this computes the maximum absolute Co and Cg values, stores the result in the blue channel, and multiplies the Co and Cg channels (0 and 1) by its reciprocal. The original Co and Cg values can then be reconstructed by multiplying by the blue channel.

The scaling information is quantized to the given number of bits.

Note

This assumes that your texture compression format uses 4x4 blocks. This is true for all BC1-BC7 formats, but ASTC can use other block sizes.

Optional Parameters

bits: Defaults to 5

Safety

The Nvidia SDK does not declare what a safe value for bits may be. If it is too large or small, it will cause undefined behaviour. Assuming the operation has not changed much since the original open sourced implementation, then bits cannot be larger than 32, as it is casted to a c_int and then performs the bitshift 1 << bits. No current version lower/upper bounds are formally stated.

Only the value bits = 5is mentioned, however a far broader range of values is likely safe.

pub fn from_ycocg(&mut self)

Converts from (Co, Cg, scale, Y) colors to (r, g, b, 1) colors.

This is useful for formats that use chroma subsampling. Y is in the range [0, 1], while Co and Cg are in the range [-1, 1]. Co and Cg are multiplied by channel 2 (scale) to reverse the effects of optionally calling Surface::block_scale_cocg().

The YCoCg-to-RGB formula used is

r = Y + Co - Cg
g = Y + Cg
b = Y - Co - Cg

pub fn to_lm(&mut self, range: Option<f32>, threshold: Option<f32>)

Stores luminance-only values in a two-channel way. Maybe consider BC4 compression instead.

Luminance L is computed by averaging the red, green, and blue values, while M stores the max of these values and threshold. The red, green, and blue channels then store L/M, and the alpha channel stores (M - threshold)/(1 - threshold).

Optional Parameters

range: Defaults to 1.0

threshold: Defaults to 0.0

pub fn to_luvw(&mut self, range: Option<f32>)

Converts from RGB colors to a (U, V, W, L) color space, much like RGBM.

All values are clamped to [0, 1]. Then a luminance-like value L is computed from RGB using

L = max(sqrtf(R^2 + G^2 + B^2), 1e-6f).

This then stores the value (R/L, G/L, B/L, L/sqrt(3)).

Optional Paramters

range: Defaults to 1.0

pub fn from_luvw(&mut self, range: Option<f32>)

Converts from Surface::to_luvw()’s color space to RGB colors.

This is the same as Surface::from_rgbm() with threshold = range * sqrt(3)).

Optional Paramters

range: Defaults to 1.0

pub fn to_log_scale(&mut self, channel: Channel, base: f32)

Replaces all values with their log with the given base.

pub fn from_log_scale(&mut self, channel: Channel, base: f32)

Inverts Surface::to_log_scale() by replacing all values x with base^x.

pub fn alpha_test_coverage(&self, alpha_ref: f32, alpha_channel: Channel) -> f32

Returns the approximate fraction (0 to 1) of the image with an alpha value greater than alpha_ref.

This function uses 8 x 8 subsampling together with linear interpolation.

Notes

alpha_ref is clamped to the range [1/256, 255/256].

pub fn scale_alpha_to_coverage(
&mut self,
coverage: f32,
alpha_ref: f32,
alpha_channel: Channel
)

Attempts to scale the alpha channel so that a fraction coverage (between 0 and 1) of the surface has an alpha greater than alpha_ref.

See also Surface::alpha_test_coverage() for the method used to determine what fraction passes the alpha test.

pub fn average(
&self,
channel: Channel,
alpha_channel: Option<Channel>,
gamma: Option<f32>
) -> f32

Computes the average of a channel, possibly with alpha or with a gamma transfer function.

If alpha_channel is None, this function computes

(sum(c[i]^gamma, i=0...numPixels)/numPixels)^(1/gamma)

where c is the channel’s data.

Otherwise, this computes

(sum((c[i]^gamma) * a[i], i=0...numPixels)/sum(a[i], i=0...numPixels))^(1/gamma)

where a is the alpha channel’s data.

Optional Parameters

gamma: Defaults to 2.2

pub fn histogram(
&self,
channel: Channel,
range_min: f32,
range_max: f32,
bins: &mut [i32]
)

Stores a histogram of channel values between range_min and range_max into bins.

This function does not clear bins’ values, in case we want to accumulate multiple histograms.

Each texel’s value is linearly mapped to a bin, using floor rounding. Values below range_min are clamped to the first bin, values above range_max are clamped to the last bin. Then the bin’s value is incremented.

Panics

Panics if bins.len() == 0.

pub fn range(
&self,
channel: Channel,
alpha_channel: Option<Channel>,
alpha_ref: f32
) -> (f32, f32)

Returns the minimum and maximum value in this channel, possibly using alpha testing.

If alpha_channel is None, this returns the minimum and maximum value in this channel.

Otherwise, this only includes texels for which the alpha value is greater than alpha_ref.

If an alpha channel is selected and all texels fail the alpha test, this sets this will return f32::MAX and to f32::MIN, i.e. for a return value ret one will have ret.0 > ret.1.

pub fn transform(
&mut self,
w0: [f32; 4],
w1: [f32; 4],
w2: [f32; 4],
w3: [f32; 4],
offset: [f32; 4]
)

Applies a 4x4 affine transformation to the RGBA values.

w0...w3 are the columns of the matrix. offset is added after the matrix-vector multiplication.

In other words, all (r, g, b, a) values are replaced with

 (r)   (w0[0], w1[0], w2[0], w3[0]) (r)   (offset[0])
 (g) = (w0[1], w1[1], w2[1], w3[1]) (g) + (offset[1])
 (b)   (w0[2], w1[2], w2[2], w3[2]) (b)   (offset[2])
 (a)   (w0[3], w1[3], w2[3], w3[3]) (a)   (offset[3])

pub fn swizzle(&mut self, r: Swizzle, g: Swizzle, b: Swizzle, a: Swizzle)

Swizzles the channels of the surface.

Each argument specifies where the corresponding channel should come from. For instance, setting r to Swizzle::B would mean that the red channel would be set to the current blue channel.

In addition, the special values Swizzle::One, Swizzle::Zero, and Swizzle::NegOne represent setting the channel to a constant value of 1.0, 0.0, or -1.0, respectively.

pub fn scale_bias(&mut self, channel: Channel, scale: f32, bias: f32)

Applies a scale and bias to the given channel. Each value x is replaced by x * scale + bias.

pub fn abs(&mut self, channel: Channel)

Replaces all colors by their absolute value.

pub fn clamp(&mut self, channel: Channel, low: f32, high: f32)

Clamps all values in the channel to the range [low, high].

pub fn blend(&mut self, rgba: [f32; 4], t: f32)

Interpolates all texels between their current color and a constant color (r, g, b, a).

t is the value used for linearly interpolating between the surface’s current colors and the constant color. For instance, a value of t=0 has no effect to the surface’s colors, and a value of t=1 replaces the surface’s colors entirely with (r, g, b, a).

pub fn convolve_slice(&mut self, channel: Channel, dim: u32, kernel: &mut [f32])

Convolves a channel with a kernel.

This uses a 2D dim x dim kernel, with values in kernel specified in row-major order. The behavior around image borders is determined by Surface::wrap_mode().

Panics

Panics if dim * dim > kernel.len()

pub fn convolve<const N: usize>(
&mut self,
channel: Channel,
kernel: [[f32; N]; N]
)

Similar to Surface::convolve_slice(), this convolves a channel with a kernel.

Values should be specified in row-major order.

pub fn binarize(&mut self, channel: Channel, threshold: f32, dither: bool)

Sets values in the given channel to either 1 or 0 depending on if they’re greater than the threshold, with optional dithering.

If dither is true, this uses Floyd-Steinberg dithering on the CPU. Not supported for 3D surfaces.

Panics

Panics if dither is true and depth > 1

pub unsafe fn quantize(
&mut self,
channel: Channel,
bits: u32,
exact_endpoints: bool,
dither: bool
)

Quantizes this channel to a particular number of bits, with optional dithering. Assumes input is in the [0, 1] range. Output is in the [0, 1] range, but rounded to the middle of each bin.

Parameters

channel: The index of the channel to quantize.

bits: The number of bits to quantize to, yielding 2^bits possible values.

exact_endpoints: If true, the set of quantized values will be 0, 1/(2^bits-1), ..., 1, rather than 0, 1/(2^bits), ..., (2^bits-1)/(2^bits).

dither: If true, uses Floyd-Steinberg dithering on the CPU. Not supported for 3D surfaces.

Panics

Panics if dither is true and depth > 1

Safety

The Nvidia SDK does not declare what a safe value for bits may be. If it is too large or small, it will cause undefined behaviour. Assuming the operation has not changed much since the original open sourced implementation, then bits cannot be larger than 32, as it is casted to a c_int and then performs the bitshift 1 << bits. No current version lower/upper bounds are formally stated.

pub fn to_normal_map(&mut self, sm: f32, medium: f32, big: f32, large: f32)

Sets the RGB channels to a normal map generated by interpreting the alpha channel as a heightmap, using a blend of four small-scale to large-scale Sobel kernels.

This uses a 9x9 kernel which is a weighted sum of a 3x3 (small), 5x5 (medium), 7x7 (big), and 9x9 (large) differentiation kernels. Each of the weights can be greater than 1, or even negative. However, the kernel will be normalized so that its elements sum to 1, so scaling should be done on the alpha channel beforehand. The smallest kernel focuses on the highest-frequency details, and larger kernels attenuate higher frequencies.

The source alpha channel, which is used as a height map to differentiate, is copied to the output.

The output RGB channels will be in the ranges [-1, 1], [-1, 1], and [0, 1].

pub fn transform_normals(&mut self, transform: NormalTransform)

Applies a 3D->2D normal transformation, setting the z (blue) channel to 0.

pub fn reconstruct_normals(&mut self, transform: NormalTransform)

Reconstructs 3D normals from 2D transformed normals.

pub fn to_clean_normal_map(&mut self)

Sets the z (blue) channel to x^2 + y^2.

If the x and y channels represent slopes, instead of normals, then this represents a CLEAN map. The important thing about this is that it can be mipmapped, and the difference between the sum of the square of the first and second mipmapped channels and the third mipmapped channel can be used to determine how rough the normal map is in a given area.

This is a lower-memory and lower-bandwidth version of LEAN mapping, but it has the drawback that it can only represent isotropic roughness.

pub fn pack_normals(&mut self)

Packs signed normals in [-1, 1] to an unsigned range [0, 1], using

(r, g, b, a) |-> (r/2+1/2, g/2+1/2, b/2+1/2, a).

pub fn unpack_normals(&mut self)

Expands packed normals in [0, 1] to signed normals in [-1, 1], using

(r, g, b, a) |-> (2r-1, 2g-1, 2b-1, a).